Add Proxy Json file to make sure the process we kill is the previous running proxy

+++ SpatialGDK/Source/SpatialGDKServices/Private/LocalReceptionistProxyServerManager.cpp
@@ -3,11 +3,17 @@
 #include "LocalReceptionistProxyServerManager.h"
 
 #include "Internationalization/Regex.h"
+#include "Misc/FileHelper.h"
+#include "Serialization/JsonReader.h"
+#include "Serialization/JsonSerializer.h"
+#include "Serialization/JsonWriter.h"
 #include "Sockets.h"
 #include "SocketSubsystem.h"
 #include "UObject/CoreNet.h"
 
 #include "SpatialCommandUtils.h"
+#include "SpatialGDKServicesConstants.h"
+#include "..\Public\LocalReceptionistProxyServerManager.h"
 
 DEFINE_LOG_CATEGORY(LogLocalReceptionistProxyServerManager);
 
@@ -15,7 +21,7 @@ DEFINE_LOG_CATEGORY(LogLocalReceptionistProxyServerManager);
 
 namespace
 {
-	static const int32 ExitCodeSuccess = 0;
+	static const FString ProxyInfoFilePath = FPaths::Combine(SpatialGDKServicesConstants::ProxyFileDirectory, SpatialGDKServicesConstants::ProxyInfoFilename);
 }
 
 FLocalReceptionistProxyServerManager::FLocalReceptionistProxyServerManager()
@@ -35,7 +41,7 @@ bool FLocalReceptionistProxyServerManager::GetProcessName(const FString& PID, FS
 	int32 ExitCode;
 	FString StdErr;
 	bSuccess = FPlatformProcess::ExecProcess(*TaskListCmd, *TaskListArgs, &ExitCode, &TaskListResult, &StdErr);
-	if (ExitCode == ExitCodeSuccess && bSuccess)
+	if (ExitCode == 0 && bSuccess)
 	{
 		FRegexPattern ProcessNamePattern(TEXT("\"(.+?)\""));
 		FRegexMatcher ProcessNameMatcher(ProcessNamePattern, TaskListResult);
@@ -85,16 +91,22 @@ bool FLocalReceptionistProxyServerManager::LocalReceptionistProxyServerPreRunChe
 	// Check if any process is blocking the receptionist port
 	if (CheckIfPortIsBound(ReceptionistPort, PID, OutLogMessage))
 	{
-		// Try killing the process that blocks the receptionist port 
-		bool bProcessKilled = SpatialCommandUtils::TryKillProcessWithPID(PID);
-		if (!bProcessKilled)
+		// Try killing the process that blocks the receptionist port if the process blocking the port is a previously running proxy.
+		if (PID == ParsePid())
 		{
-			UE_LOG(LogLocalReceptionistProxyServerManager, Warning, TEXT("%s %s!"), *LOCTEXT("FailedToKillBlockingPortProcess", "Failed to kill the process that is blocking the port.").ToString(),*OutLogMessage.ToString());
-			return false;
+			bool bProcessKilled = SpatialCommandUtils::TryKillProcessWithPID(PID);
+			if (!bProcessKilled)
+			{
+				UE_LOG(LogLocalReceptionistProxyServerManager, Warning, TEXT("%s %s!"), *LOCTEXT("FailedToKillBlockingPortProcess", "Failed to kill the process that is blocking the port.").ToString(), *OutLogMessage.ToString());
+				return false;
+			}
+
+			UE_LOG(LogLocalReceptionistProxyServerManager, Log, TEXT("%s %s."), *LOCTEXT("KilledBlockingPortProcess", "Succesfully killed").ToString(), *OutLogMessage.ToString());
+			return true;
 		}
 
-		UE_LOG(LogLocalReceptionistProxyServerManager, Log, TEXT("%s %s."), *LOCTEXT("KilledBlockingPortProcess", "Succesfully killed").ToString(), *OutLogMessage.ToString());
-		return true;
+		UE_LOG(LogLocalReceptionistProxyServerManager, Log, TEXT("The required port is blocked from a different process with Pid: %s"), *PID);
+		return false;
 	}
 
 	OutLogMessage = LOCTEXT("NoProcessBlockingPort", "The required port is not blocked!");
@@ -124,6 +136,80 @@ bool FLocalReceptionistProxyServerManager::TryStopReceptionistProxyServer()
 }
 
 
+TSharedPtr<FJsonObject> FLocalReceptionistProxyServerManager::ParsePidFile()
+{
+	FString ProxyInfoFileResult;
+	TSharedPtr<FJsonObject> JsonParsedProxyInfoFile;
+
+	if (FFileHelper::LoadFileToString(ProxyInfoFileResult, *ProxyInfoFilePath))
+	{
+		if (FSpatialGDKServicesModule::ParseJson(ProxyInfoFileResult, JsonParsedProxyInfoFile))
+		{
+			return JsonParsedProxyInfoFile;
+		}
+		else
+		{
+			UE_LOG(LogLocalReceptionistProxyServerManager, Error, TEXT("Json parsing of proxyInfo.json failed. Can't get proxy's Pid."));
+		}
+	}
+	else
+	{
+		UE_LOG(LogLocalReceptionistProxyServerManager, Error, TEXT("Loading proxyInfo.json failed. Can't get proxy's Pid."));
+	}
+
+	return nullptr;
+}
+
+void FLocalReceptionistProxyServerManager::SetPidInJson(const FString& Pid)
+{
+	FString ProxyInfoFileResult;
+
+	// If file does not exist create an empty json file 
+	if (!FPaths::FileExists(ProxyInfoFilePath))
+	{
+		FFileHelper::SaveStringToFile(TEXT("{ }"), *ProxyInfoFilePath);
+	}
+
+	TSharedPtr<FJsonObject> JsonParsedProxyInfoFile = ParsePidFile();
+	if (!JsonParsedProxyInfoFile.IsValid())
+	{
+		UE_LOG(LogLocalReceptionistProxyServerManager, Error, TEXT("Failed to update Pid(%s). Please ensure that the following file exists: %s"), *Pid, *SpatialGDKServicesConstants::ProxyInfoFilename);
+		return;
+	}
+
+	JsonParsedProxyInfoFile->SetStringField("pid", Pid);
+
+	TSharedRef<TJsonWriter<>> JsonWriter = TJsonWriterFactory<>::Create(&ProxyInfoFileResult);
+	if (!FJsonSerializer::Serialize(JsonParsedProxyInfoFile.ToSharedRef(), JsonWriter))
+	{
+		UE_LOG(LogLocalReceptionistProxyServerManager, Error, TEXT("Failed to write pid to parsed proxy info file. Unable to serialize content to json file."));
+		return;
+	}
+	if (!FFileHelper::SaveStringToFile(ProxyInfoFileResult, *FPaths::Combine(SpatialGDKServicesConstants::ProxyFileDirectory, SpatialGDKServicesConstants::ProxyInfoFilename)))
+	{
+		UE_LOG(LogLocalReceptionistProxyServerManager, Error, TEXT("Failed to write file content to %s"), *SpatialGDKServicesConstants::ProxyInfoFilename);
+	}
+}
+
+FString FLocalReceptionistProxyServerManager::ParsePid()
+{
+	FString Pid;
+
+	if (TSharedPtr<FJsonObject> JsonParsedProxyInfoFile = ParsePidFile())
+	{
+		if (JsonParsedProxyInfoFile->TryGetStringField(TEXT("pid"), Pid))
+		{
+			return Pid;
+		}
+		else
+		{
+			UE_LOG(LogLocalReceptionistProxyServerManager, Error, TEXT("'pid' does not exist in proxyInfo.json. Can't read proxy's Pid."));
+		}
+	}
+
+	Pid.Empty();
+	return Pid;
+}
 
 bool FLocalReceptionistProxyServerManager::TryStartReceptionistProxyServer(bool bIsRunningInChina, const FString& CloudDeploymentName, const FString& ListeningAddress, const int32 ReceptionistPort)
 {
@@ -159,6 +245,15 @@ bool FLocalReceptionistProxyServerManager::TryStartReceptionistProxyServer(bool 
 		const FText WarningMessage = FText::Format(LOCTEXT("FailedToStartProxyServer", "Starting the local receptionist proxy server failed. Error Code: {0}, Error Message: {1}"), ExitCode, FText::FromString(StartResult));
 		UE_LOG(LogLocalReceptionistProxyServerManager, Warning, TEXT("%s"), *WarningMessage.ToString());
 		return false;
+	}
+
+	FString Pid;
+	FString State;
+
+	// Save the server receptionist proxy process's Pid in a Json file
+	if (SpatialCommandUtils::GetProcessInfoFromPort(ReceptionistPort, Pid, State))
+	{
+		SetPidInJson(Pid);
 	}
 
 	RunningCloudDeploymentName = CloudDeploymentName;
